updated server code to use arangodb
