+++ arangodb.log
@@ -1,0 +1,7 @@
+2020-10-20T10:12:44-07:00 |INFO| Starting arangodb version 0.14.15, build e32307e component=arangodb
+2020-10-20T10:12:44-07:00 |INFO| Using storage engine 'rocksdb' component=arangodb
+2020-10-20T10:12:44-07:00 |INFO| Serving as master with ID '3d8c5a02' on :8528... component=arangodb
+2020-10-20T10:12:44-07:00 |INFO| Waiting for 3 servers to show up.
+ component=arangodb
+2020-10-20T10:12:44-07:00 |INFO| Use the following commands to start other servers: component=arangodb
+2020-10-20T10:12:44-07:00 |INFO| ArangoDB Starter listening on 0.0.0.0:8528 (:8528) component=arangodb
