+++ client/src/app.jsx
@@ -13,7 +13,7 @@ function App() {
   // eslint-disable-next-line camelcase
   function getImages(listingId) {
     axios.get(`/api/images/${listingId}`)
-      .then(({ data }) => { setImageData(data); })
+      .then(({ data }) => { console.log(data); setImageData(data); })
       .catch((err) => { console.log('Error getting image data', err); });
   }
