+++ package.json
@@ -17,6 +17,7 @@
   },
   "dependencies": {
     "@types/jest": "^26.0.14",
+    "arangojs": "^5.8.0",
     "axios": "^0.20.0",
     "body-parser": "^1.19.0",
     "compression": "^1.7.4",
